refactor(Operator) Entanglement as a separate class
